Rename couchbase configuration, minor cleanup in tests
